Move breakpoints to utils dir
