+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { cleanHTML } from "./cleanHTML";
+export { breakpoints } from "./breakpoints";
